+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## next [????.??.??]
+* Introduce the `Data.Eliminator.TH` module, which provides functionality for
+  generating eliminator functions using Template Haskell. Currently, only
+  simple algebraic data types that do not use polymorphic recursion are
+  supported.
 * All eliminators now use predicates with `(~>)`.
 
 ## 0.1 [2017-07-02]